// src/shared/api/models/putApiEconomicConfigBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiEconomicConfigBodyCoaConfiguration } from './putApiEconomicConfigBodyCoaConfiguration';

export type PutApiEconomicConfigBody = {
    /** e-conomic organization ID */
    organizationId?: string;
    /** Chart of accounts configuration */
    coaConfiguration?: PutApiEconomicConfigBodyCoaConfiguration;
    /** Whether to send emails from e-conomic */
    sendEmail?: boolean;
};